Replace: AuthModal to AuthScreen
