+++ src/pages/AuthScreen/styles.tsx
