+++ src/pages/HomeScreen/index.tsx
@@ -7,7 +7,6 @@ import ViewTitle from '../../components/ViewTitle';
 import ShipmentList from '../../components/ShipmentList';
 import {useEffect, useState} from 'react';
 import useAPI from '../../hooks/useAPI';
-import AuthModal from '../../components/AuthModal';
 import NewShipmentButton from '../../components/NewShipmentButton';
 import ShipmentModal from '../../components/ShipmentModal';
 
@@ -15,7 +14,6 @@ const HomeScreen = () => {
   const {welcome} = useAPI();
   const [show, setShow] = useState<boolean>(false);
   const isDarkMode = useColorScheme() === 'dark';
-
   const backgroundStyle = {
     backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
   };
@@ -32,7 +30,6 @@ const HomeScreen = () => {
       contentInsetAdjustmentBehavior="automatic"
       style={backgroundStyle}>
       <Section>
-        <AuthModal />
         <MainHeader />
         <ViewTitle value="Home - Delivery List" />
         <NewShipmentButton onPress={openModal} />
