+++ src/pages/index.tsx
@@ -1,6 +1,7 @@
 import {NavigationContainer} from '@react-navigation/native';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import ApiProvider from '../contexts/ApiContext';
+import AuthScreen from './AuthScreen';
 import DetailScreen from './DetailScreen';
 import HomeScreen from './HomeScreen';
 
@@ -10,7 +11,13 @@ const MainRouter = () => {
   return (
     <ApiProvider>
       <NavigationContainer>
-        <Stack.Navigator initialRouteName="Home">
+        <Stack.Navigator initialRouteName="Auth">
+          <Stack.Screen
+            name="Auth"
+            component={AuthScreen}
+            options={{headerShown: false}}
+          />
+
           <Stack.Screen
             name="Home"
             component={HomeScreen}
